+++ RazzleServer/Map/MapleSummon.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using System.Threading;
+using MapleLib.PacketLib;
 
 namespace RazzleServer.Map
 {
@@ -69,7 +70,7 @@ namespace RazzleServer.Map
         public PacketWriter GetSpawnPacket(bool spawnAnimated)
         {
             
-            var pw = new PacketWriter(SMSGHeader.SPAWN_SPECIAL_MAPOBJECT); // SPAWN_SUMMON
+            var pw = new PacketWriter((ushort)SMSGHeader.SPAWN_SPECIAL_MAPOBJECT); // SPAWN_SUMMON
 
             pw.WriteInt(Owner.ID);
             pw.WriteInt(ObjectID);
@@ -103,7 +104,7 @@ namespace RazzleServer.Map
         public PacketWriter RemovePacket(bool animated)
         {
             
-            var pw = new PacketWriter(SMSGHeader.REMOVE_SPECIAL_MAPOBJECT); // REMOVE_SUMMON
+            var pw = new PacketWriter((ushort)SMSGHeader.REMOVE_SPECIAL_MAPOBJECT); // REMOVE_SUMMON
 
             pw.WriteInt(Owner.ID);
             pw.WriteInt(ObjectID);
@@ -118,7 +119,7 @@ namespace RazzleServer.Map
         public PacketWriter MovePacket(Point startPosition, List<MapleMovementFragment> movementList)
         {
             
-            var pw = new PacketWriter(SMSGHeader.MOVE_SUMMON);
+            var pw = new PacketWriter((ushort)SMSGHeader.MOVE_SUMMON);
 
             pw.WriteInt(Owner.ID);
             pw.WriteInt(ObjectID);
@@ -134,7 +135,7 @@ namespace RazzleServer.Map
         public PacketWriter GetUseSkillPacket(int skillId, byte stance)
         {
             
-            var pw = new PacketWriter(SMSGHeader.SUMMON_SKILL);
+            var pw = new PacketWriter((ushort)SMSGHeader.SUMMON_SKILL);
             pw.WriteInt(Owner.ID);
             pw.WriteInt(skillId);
             pw.WriteByte(stance);
@@ -144,7 +145,7 @@ namespace RazzleServer.Map
         public PacketWriter GetAttackPacket(AttackInfo attackInfo, bool darkFlare = false)
         {
             
-            var pw = new PacketWriter(SMSGHeader.SUMMON_ATTACK);
+            var pw = new PacketWriter((ushort)SMSGHeader.SUMMON_ATTACK);
             pw.WriteInt(Owner.ID);
             pw.WriteInt(ObjectID);
             pw.WriteByte(Owner.Level);
